service-workers/service-worker/ServiceWorkerGlobalScope/service-worker-error-event.https.html

<!DOCTYPE html>
<title>ServiceWorkerGlobalScope: Error event error message</title>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='../resources/test-helpers.sub.js'></script>
<script>
promise_test(async t => {
  const SCRIPT = 'resources/error-worker.js';
  const SCOPE = 'resources/scope/service-worker-error-event';
  const registration = await service_worker_unregister_and_register(
    t,
    SCRIPT,
    SCOPE
  );
  const worker = registration.installing;
  const event = await new Promise(function(resolve) {
    navigator.serviceWorker.onmessage = resolve;
    worker.postMessage('');
  });

  add_completion_callback(async() => {
    if (registration)
      await registration.unregister();
  });

  assert_equals(event.data.error, 'testError', 'error type');
  assert_greater_than(
    event.data.message.indexOf('testError'), -1, 'error message');
  assert_greater_than(
    event.data.filename.indexOf(SCRIPT), -1, 'filename');
  assert_equals(event.data.lineno, 5, 'error line number');
  assert_equals(event.data.colno, 3, 'error column number');

}, 'Error handlers inside serviceworker should see the attributes of ' +
   'ErrorEvent');
</script>
