Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/ServiceWorkerGlobalScope/service-worker-error-event.https-expected.txt


FAIL Error handlers inside serviceworker should see the attributes of ErrorEvent assert_equals: error column number expected 3 but got 20
